+++ src/components/pages/RootPage.jsx
@@ -15,7 +15,7 @@ import CreatePageContainer from '../../containers/pages/CreatePageContainer';
 import DashboardPage from './DashboardPage';
 import ExportViewerPageContainer from '../../containers/pages/ExportViewerPageContainer';
 import ToolPageContainer from '../../containers/pages/ToolPageContainer';
-import LoginPage from './LoginPage';
+import WelcomePage from './WelcomePage';
 import LogoutPageContainer from '../../containers/pages/LogoutPageContainer';
 import RecordPageContainer from '../../containers/pages/RecordPageContainer';
 import ReportViewerPageContainer from '../../containers/pages/ReportViewerPageContainer';
@@ -62,7 +62,7 @@ function RootPage(props) {
       </Helmet>
 
       <Switch>
-        <PublicRoute path="/login" component={LoginPage} />
+        <PublicRoute path="/login" component={WelcomePage} />
         <PublicRoute path="/logout" component={LogoutPageContainer} decorated={false} />
         <PublicRoute path="/config" component={ConfigPage} />
         <PublicRoute path="/authorize" component={AuthorizePageContainer} decorated={false} />
